requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.6.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
beautifulsoup4==4.13.4
chardet==5.2.0
